requirements.txt
opentelemetry.instrumentation.flask==0.45b0
opentelemetry.instrumentation.django==0.45b0
opentelemetry.instrumentation.logging==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-openai==0.19.0
opentelemetry-instrumentation-chromadb==0.19.0
python-dotenv==1.0.0